move common.warnings
