+++ src/neptune/internal/operation_processors/read_only_operation_processor.py
@@ -17,11 +17,11 @@ __all__ = ("ReadOnlyOperationProcessor",)
 
 from typing import TYPE_CHECKING
 
-from neptune.common.warnings import (
+from neptune.internal.operation_processors.operation_processor import OperationProcessor
+from neptune.internal.warnings import (
     NeptuneWarning,
     warn_once,
 )
-from neptune.internal.operation_processors.operation_processor import OperationProcessor
 
 if TYPE_CHECKING:
     from neptune.internal.operation import Operation
