+++ src/neptune/internal/utils/deprecation.py
@@ -16,8 +16,8 @@
 from functools import wraps
 from typing import Optional
 
-from neptune.common.warnings import warn_once
 from neptune.exceptions import NeptuneParametersCollision
+from neptune.internal.warnings import warn_once
 
 __all__ = ["deprecated", "deprecated_parameter"]
