+++ src/neptune/metadata_containers/metadata_container.py
@@ -44,7 +44,6 @@ from neptune.attributes.attribute import Attribute
 from neptune.attributes.namespace import Namespace as NamespaceAttr
 from neptune.attributes.namespace import NamespaceBuilder
 from neptune.common.utils import reset_internal_ssl_state
-from neptune.common.warnings import warn_about_unsupported_type
 from neptune.envs import (
     NEPTUNE_ENABLE_DEFAULT_ASYNC_LAG_CALLBACK,
     NEPTUNE_ENABLE_DEFAULT_ASYNC_NO_PROGRESS_CALLBACK,
@@ -93,6 +92,7 @@ from neptune.internal.utils.logger import (
 from neptune.internal.utils.paths import parse_path
 from neptune.internal.utils.uncaught_exception_handler import instance as uncaught_exception_handler
 from neptune.internal.value_to_attribute_visitor import ValueToAttributeVisitor
+from neptune.internal.warnings import warn_about_unsupported_type
 from neptune.metadata_containers.abstract import (
     NeptuneObject,
     NeptuneObjectCallback,
