+++ tests/unit/neptune/new/client/test_project.py
@@ -23,10 +23,6 @@ from neptune import (
     ANONYMOUS_API_TOKEN,
     init_project,
 )
-from neptune.common.warnings import (
-    NeptuneWarning,
-    warned_once,
-)
 from neptune.envs import (
     API_TOKEN_ENV_NAME,
     PROJECT_ENV_NAME,
@@ -41,6 +37,10 @@ from neptune.internal.backends.api_model import (
 )
 from neptune.internal.backends.neptune_backend_mock import NeptuneBackendMock
 from neptune.internal.exceptions import NeptuneException
+from neptune.internal.warnings import (
+    NeptuneWarning,
+    warned_once,
+)
 from neptune.metadata_containers.utils import (
     DATE_FORMAT,
     parse_dates,
